project-trade_support/requirements.txt
ccxt>=4.0.0
pandas>=2.0.0
numpy>=1.24.0
openai>=1.0.0
anthropic>=0.20.0
google-generativeai>=0.5.0
streamlit>=1.30.0
apscheduler>=3.10.0
python-dotenv>=1.0.0
plotly>=5.18.0
requests>=2.31.0
ta>=0.11.0
